backend/requirements.txt
# python=3.12pytho
python-dotenv
requests
supabase